[iOS] Fix omnibox row swipe-to-delete layout
In the omnibox row, the leading icon and the text are positioned
relative to the layout guide on top of the omnibox. Unfortunately, when
you swipe a row to delete it, this leads to the icon and text not
moving because the corresponding omnibox icons don't move.

To fix this, we freeze the position of the icon and text relative to
the contentView when the cell enters the deletion state and unfreeze
when the deletion ends.

Bug: 1063343
Change-Id: I023267d47b7bfc33137ad71049df4234d311d6b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2116172
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#752882}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a86c63a5c732452f52b227c1ef9cdc288707f7f7