[iOS] Remove direct BVC dependencies in MainController.
Removes direct use of the BVC in MainController. The most common uses of
the BVC through the interface provider are (a) access to the dispatcher,
and (b) marking the BVC as the primary one.

(This CL doesn't address whether the notion of a BVC as "primary" is
still useful).

The dispatcher can be accessed through an interface's Browser.

This CL adds the -setPrimary: API to BrowserInterface for cases where
it is needed.

Additionally, updates removeBrowsingDataForBrowserState: to handle --
possibly correctly -- multiple connected scenes. This just repeats all
of the setInteractionEnabled/showActivityOverlay calls for each scene.

Change-Id: Ia6bebf86a991a977f83f3f3a8da086fe5de519e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2124447
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#754001}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bb73db5761315ed32a82205bc98893bb95b0dc45