[iOS] Fix NTP visibility for cancelled navigations
When a navigation is cancelled, the NTP should still be presented.
Currently, as the navigation starts, the NTP is hidden and it is not
restored when the navigation is cancelled.

Fixed: 1063154
Change-Id: I0844fa735ae4e0f9e184d46ac6c96e809b873807
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2120352
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Reviewed-by: David Jean <djean@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#754010}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eb3847fdb223fe9be5d56c7aab520a919cfe6b1e