Prepend alt text to context menu title if element is an image without a link.
When long-pressing an image element that is not a link, context menu
title should be "[<alt-text> – ]<src>". Alt text was never used before
so it is now retrieved in the js to make this possible. A function was
created to get the context menu title since the logic was too coupled
with the context menu dictionary function.

Bug: 1064569
Change-Id: I726b8066116426df3d51e89a65f906f77d739900
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2122752
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Adam Trudeau-Arcaro <adamta@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#754666}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3f44bf3ef1932b258a95ae04b007b42a78888199