[iOS] Remove some spurious dependencies on BVC through browserInterface
In order to remove explicit dependencies on the BVC, this CL takes the
initial step of using the |viewController| property in BrowserInterface
instead of |bvc| wherever possible. It also uses the Browser's command
dispatcher (through HandlerForProtocol or a static cast) instead of the
BVC's |dispatcher| property.

Remaining uses of the BVC property of browserInterface after this CL
are limited to (a) SceneController, and (b) BVC-specific public API,
which will be factored away in future CLs.

As related cleanup, the settings nav controller delegate property is
renamed from "dispatcherForSettings" to "handlerForSettings".

Change-Id: I1a10e7b1655f22f6559249597a6f4f467ed9ed87
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2124668
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#754952}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 95761677a1ee8073d08fd958277ae87d47681d6c