[ios fuzzing] Add a script to launch WebKitTestRunner on fuzzer bots
This adds a run-clusterfuzz script that can be used to launch
WebKitTestRunner on ClusterFuzz bots. Checking this script into the
repository will make it easier for buildbots to package the script along
with the rest of the build when creating builds for fuzzing.

Bug: 1053304
Change-Id: Ie2e311b862265c008e518b7d00b609265976ec5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2140034
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Ali Juma <ajuma@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#757189}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 82620847442f24bb262cdf7103c2a6b275e5a324