[iOS] Show app launcher dialog in opener for new tabs.
If an app launcher navigation occurs in a new tab, the tab is closed
immediately after the WebPolicyDecider fails.  This means that the
OverlayRequest for that dialog is cancelled before it gets a chance to
be shown.  This did not occur in the legacy implementation of this
dialog since that implementation is tied to the BVC rather than to a
particular WebState (the legacy implementation is now broken as well
due to some recent refactoring that nils out the AppLauncherCoordinator
base view controller).

This CL updates the coordinator to enqueue the request for the app
launcher dialog into the OverlayRequestQueue for the tab's opener if
there the new tab is going to be closed immediately after.

Bug: 1041867
Change-Id: I2d6b94a615d2bed0d8e475c6fdbbf428912ffd5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2137822
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#757303}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7ba526ebf50f92a25a2db081a982736d221e8782