[iOS] Conflict between Sign-in Promo and FRE
This case should never happens for regular users since the sign-in
promo would never be triggered when Chrome is started for the first time.
This fix is only for debug reasons.
The -[ScheneController scheduleShowPromo] needs to be called after
triggering the first run. This is to avoid starting both, the FRE and
the sign-in promo.

The bug is, |scheduleShowPromo| is called by
|startUpChromeUIPostCrash:needRestoration:| before |showFirstRunUI|

Change-Id: Idb0e32fdd30f7fb72165f2c8bdfb6a7571bdd121
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2139824
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#757425}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 17a1248b393e8751c884af9e4a858b61b1de150e