[ios] Add ManagedBookmark flags
This CL creates two flags:

1) ShouldInstallManagedBookmarksPolicyHandler() checks the presence of
a command line flag to decide whether to install the ManagedBookmarks
policy handler. See below for explanation of the need to use a command
line flag.

2) IsManagedBookmarksEnabled() checks a Finch feature flag to decide
whether to show the managed bookmarks in the bookmarks UI. A Finch
feature flag is used to act as a server-side kill switch.


The policy system is initialized before about:flags or field trials, so
it is not possible for either of those mechanisms to decide whether or
not policy should be enabled at runtime. Instead, we will switch to
looking for the presence (or absence) of a command line flag. This will
allow us to easily turn policy on locally during development, and when
we are ready to ship policy support, we can remove the command line
flag.

Bug: 1065187
Change-Id: I1aea8c106642ec86b422e4bb333294d05d3dba95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2142752
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#757906}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 68696843b15c1c4dd2d18ccd25b1ce96902185ba