[iOS] Update WebStatePolicyDecider::ShouldAllowRequest return value
Replace the boolean return value of
WebStatePolicyDecider::ShouldAllowRequest with a struct PolicyDecision.

PolicyDecision will be extended in a follow-up CL to specify an error
which will denote the specific failure of loading a given
|request|. Additionally, PolicyDecision::RequestAction is an enum as
it will also be extended with an additional rejection type to reject the
load and present UI displaying an error page with details about the
given error.

Bug: 1065161
Change-Id: I4e7eabf7ed4055bbd1365dab05bda53b5aaf2e1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2130971
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#757955}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 15f6db599b9b10e28c6c726fd1351d15219392f4