[ios] Temporary fix for CRWSessionStorage being released early in ClosingWebStateObserver.
RestoreIOSLiveTab is not compiled with ARC and therefore is not properly
retaining the CRWSessionStorage* which it is passed. Temporarily work
around this bug by keeping a strong pointer to the CRWSessionStorage* in
ClosingWebStateObserver, until RestoreIOSLiveTab can be fixed properly.

BUG=1070852

Change-Id: Ic3d8a4d0cc3c29bfae4f657de06b98aa869b3bc1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2150035
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#759116}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 82d5a0c79ebaaa84af40e2eae494fe596773f7f0