[iOS] Block the content area for HTTP auth dialogs from a different host
This CL refactors the browser container to use a mediator and consumer.
The BrowserContainerMediator observes presentation of HTTP
authentication dialogs.  If one is presented from a different host than
the rendered page in the content area, a blocking view is inserted
behind the HTTP auth overlay to block out the page.

In order to support this change, a URL is added to the OverlayRequest
config so that it can be compared with the last committed URL of the
Browser's active WebState.

Bug: 1063690
Change-Id: I21c7abe83818180762e3a2dda6f764ea635ee42a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2136232
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#759367}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0068e74c4235b549731b61e889ee502602c168fb