[ios] Add EditBookmarks experiment flag
EditBookmarksEnabled is an enterprise policy that determines whether
the user can edit any bookmarks. The policy is mapped to a user pref.

This CL is part of a series of CLs that add support for this
enterprise policy on the iOS platform.

Bug: 1072326
Change-Id: I6fd2e640e35191b4e6feec6fd41b9e063c7f2a83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2153772
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#760476}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 39d08d618ad35f9dd8f22d9bd784d124fa34fb0f