[ios] Add EditBookmarksEnabled enterprise policy handler
EditBookmarksEnabled is an enterprise policy that determines whether
the user can edit any bookmarks. The policy is mapped to a user pref.

This CL is part of a series of CLs that add support for this
enterprise policy on the iOS platform.

This CL:
- Turns on iOS support for EditBookmarksEnabled.
- Adds EditBookmarksEnabled to basic policy handler policy map.
- Adds basic test via json test cases.

Bug: 1072326
Change-Id: I1b03db4ee2d808bf5c4c7bbe039cf3346d98c9b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2156138
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: Owen Min <zmin@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#760600}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9579c164ea1c5a4a35cc169cb013a64c0559cb8b