[ios] Handle EditBookmarks pref in ui/toolbar/
EditBookmarksEnabled is an enterprise policy that determines whether
the user can edit any bookmarks. The policy is mapped to a user pref.

This is part of a series of CLs that add support for this
enterprise policy on the iOS platform.

Specifically in this CL, the add/edit bookmark button on the tablet
toolbar is disabled when EditBookmarksEnabled is false.

Bug: 1072326
Change-Id: I646558b118e759fb22035c7f2455c62e5aef7232
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2156144
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#760825}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: db4268e62a5dfc6afcd17abae9ce7464ca10cd51