[iOS] Prevent navigations in one tab from closing FIP in another tab
The underlying issue is that the find_bar_mediator was observing
navigation events and closing FIP when a navigation occurs. As the
state for "show FIP" is a tab-level state, the observation should occur
in the tab helper.

Now, the tab helper observes the navigation in its own webstate.
Alerting the UI that Find In Page has stopped is done through the
FindInPageResponseDelegate. (This used to be the coordinator, but
should really be the mediator, so that has changed too.) This is hooked
up to the active mediator only when the UI is active.

From the bug, the same issue occurs with Text Zoom, but Text Zoom is
closed automatically on switching tabs, so that behavior was kept.

Fixed: 1067603
Change-Id: Ic5b1fb61142ab3deadc8ef792642e880cf23ebea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2157226
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#760945}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5f73d42d5b5ade652751feed5c43544e14fe0ae4