[ios] Observe EditBookmarksEnabled pref changes in ui/toolbar
EditBookmarksEnabled is an enterprise policy that determines whether
the user can edit any bookmarks. The policy is mapped to a user pref.

This is part of a series of CLs that add support for this
enterprise policy on the iOS platform.

Specifically in this CL, the add/edit bookmark button on the tablet
toolbar is dynamically updated when EditBookmarksEnabled pref is
changed.

Bug: 1072326
Change-Id: If076ed307e9d9d4ee3e3eabce0411047440ae391
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2158834
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#761188}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 687dc0f4520cadd3259e3ab39189e471899601f4