[ios] Add ManagedBookmarksEnabled pref observer
ManagedBookmarksEnabled is an enterprise policy that shows a permanent
"Managed Bookmarks" folder as a sibling of the Mobile Bookmarks folder.
The policy is mapped to a corresponding user pref.

This is part of a series of CLs that add support for this
enterprise policy on the iOS platform.

Specifically in this CL:
- Add ManagedBookmarksEnabled pref observer so the bookmarks UI is
updated dynamically when the pref value is updated.

There are no existing unit tests for bookmark_home_mediator. Eg tests
will be added in a followup CL.

Bug: 1065187
Change-Id: I094bb427d0f1d5ea009e8ecd56b8813a652b7ebe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2159423
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#761264}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 87541e2f53054241a8b739b3e9e40c80cccdab3c