[ios] Handle EditBookmarks pref in ui/popup_menu/
EditBookmarksEnabled is an enterprise policy that determines whether
the user can edit any bookmarks. The policy is mapped to a user pref.

This CL is part of a series of CLs that add support for this
enterprise policy on the iOS platform.

Specifically in this CL, the add/edit bookmark tools menu item is
disabled when EditBookmarksEnabled is false.

Bug: 1072326
Change-Id: I3e9ce72e887518503867e661ecffacc076612ad8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2156142
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#761879}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d14d3ce7077caca1dda7c2bb9ee62dd2e4405fc7