[iOS][Passwords] Get dispatcher via |browser|
A recent change on browser view controller stop setting the dispatcher
on Password Controller. This is preventing PasswordBreachCommands to be
dispatched, as well as setting a nil dispatcher in the delegate in
IOSChromeSavePasswordInfoBarDelegate.

The fix in this CL is to get the dispatcher from the new passed object
|browser| and cast it using HandlerForProtocol. Also adds some DCHECKs
to try to catch this issue early next time.

Cleans up now unused dispatcher property.

Bug: 1075961
Change-Id: Iaef1df9870b9b7f788dcf5175d60ded86bd9eb90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2170095
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#763328}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 272fabc72409d20cc2dbf2028424126d2b6edddb