[iOS] UrlCheckerDelegateImpl::StartDisplayingBlockingPageHelper() impl
Implements StartDisplayingBlockingPageHelper() for the iOS
UrlCheckerDelegate.  On content//, this call is forwarded to the
SafeBrowsingService's UI manager.  However, with the committed
interstitial behavior, the UI manager class is largely just used to
implement the whitelist and unsafe resource copying functionality,
which is implemented by SafeBrowsingUrlAllowList and
SafeBrowsingUnsafeResourceStack on iOS.

This CL updates UrlCheckerDelegateImpl to execute the UnsafeResource
callbacks and modify the allow list and unsafe resource stack.  After
this fix, the SafeBrowsingTabHelper will receive do-not-proceed
signals from the URL checker and can use them to cancel unsafe
navigations.

A follow-up CL will use the information stored in the unsafe
resource stack to populate the error page shown for the cancelled
navigations.

This CL also adds the RunUnsafeResourceCallback() utility function
which executes the unsafe resource's callback on the appropriate
thread if it is non-null.

Bug: 1064803
Change-Id: I22ec557785799452cc3a026a088c9a3863617623
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2167005
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#763607}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6fcc2441f3da03455396935692aa2008e5e882e1