[iOS] Convert JavaScript dialog overlay to use Alert[Request/Response]
This CL updates the JS dialog to use the new AlertRequest and
AlertResponse classes.  This allows for the removal the mediators for
the three dialog types.

Additionally, this CL combines the request/response types for all
JavaScript dialogs into a single class, allowing the removal for the
separate requests/responses for alerts, confirmations, and prompts.
JavaScriptDialogSource was also removed since it no longer needs to
be shared among three request types.  Additionally, the new
web::WebState::Getter callback type can be used in lieu of keeping a
WebState pointer in the source that was reset to nullptr when the
WebState was destroyed.

Moreover, this CL updates OverlayJavaScriptDialogPresenter to integrate
these changes.  It updates the metrics collection to account for the
fact that alerts did not send a response previously (they were
mistakenly regarded as tab closures and navigations).

Bug: 1071543
Change-Id: I33c142ad0088edf1582c20baf9d08eadfe60788e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2155478
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#763662}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 49b320bf557d7e0b70d7b37c048128e24f5348e5