[iOS] Hide the confirmation button when the sign-in is in progress
When the sign-in is in progress, all the UI should be blocked (the
confirm button, the link and the account picker).
When the sign-in is canceled, the UI should be unblocked.

Also the activity indicator needs to be triggered even when the
sign-in is started by tapping on the setting link.

Adding:
-[UserSigninViewController signinWillStart]
-[UserSigninViewController signinDidStop]

Small details, the activity indicator is moved from the center of
the screen to the same position than the confirmation button.

Before:
https://drive.google.com/open?id=1Gaxsy-jMvbhqNGiOVspdZHGTyPjvp-rH
https://drive.google.com/open?id=1AGCO2k9ERaMNMrJYwSZNMRFBT8Q8LxWH

After:
https://drive.google.com/open?id=1WoIuHN21-TDic72b3wdYqV_rwVKpN-dD
https://drive.google.com/open?id=1uy9GaTXJ39B__sARksMB86sLdCbi6k21

Bug: 971989
Change-Id: I254ffb088862c8373a31ef00aa294b194d3add69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2172788
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#764338}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e4c427040820372e0e405c88e94379fd03d54b0d