[iOS] Update IOSBlockingPageTabHelper to support CancelAndDisplayError()
For committed interstitials shown due to CancelAndDisplayError() for
a disallowed request, the resulting navigation is committed before the
blocking page is created and associated with the tab helper.  As a
result, the previous implementation of the blocking page tab helper
where the blocking page is stored until the navigation is committed
will not work.  This CL updates the tab helper to store the last-
committed navigation ID so that calls to AssociateBlockingPage() that
occur after the navigation is committed correctly uses the blocking
page for the committed page.

In addition, this CL updates the tab helper to be created in
AttachTabHelpers() because the lazy instantiation that was occurring
in the static AssociateBlockingPage() function created the tab helper
too late to record the navigation ID of the committed page that prompts
showing the blocking page.  As a side effect, AssociateBlockingPage()
was converted into an instance method that is called on the tab helper.

Finally, this CL updates the tab helper implementation to use a nested
WebStateObserver to listen for the changing navigation ID and updates
it to use a ScopedObserver.  Moving the WebStateObserver to a private
class ensures that it will only ever be added to the tab helper's
WebState, allowing for the removal of the |web_state_| ivar and the
DCHECKs verifying that the observer callbacks are being executed for
the expected WebState.

Bug: 1064803
Change-Id: I0cf37c0bb7b46afbacf4507120e981443815c14e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173751
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Livvie Lin <livvielin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#764483}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2b376c0b15124c27710a4a2bb8e45e52911109d1